Fix test for 32-bit

+++ tests/processor/fingerprint_test.cpp
@@ -40,7 +40,7 @@ TEST(TestHttpEndpointFingerprint, Basic)
     EXPECT_EQ(output.type, DDWAF_OBJ_STRING);
     EXPECT_EQ(attr, object_store::attribute::none);
 
-    std::string_view output_sv{output.stringValue, output.nbEntries};
+    std::string_view output_sv{output.stringValue, static_cast<std::size_t>(static_cast<std::size_t>(output.nbEntries))};
     EXPECT_STRV(output_sv, "http-get-0ede9e60-0ac3796a-9798c0e4");
 
     ddwaf_object_free(&query);
@@ -71,7 +71,7 @@ TEST(TestHttpEndpointFingerprint, EmptyQuery)
     EXPECT_EQ(output.type, DDWAF_OBJ_STRING);
     EXPECT_EQ(attr, object_store::attribute::none);
 
-    std::string_view output_sv{output.stringValue, output.nbEntries};
+    std::string_view output_sv{output.stringValue, static_cast<std::size_t>(output.nbEntries)};
     EXPECT_STRV(output_sv, "http-get-0ede9e60--9798c0e4");
 
     ddwaf_object_free(&query);
@@ -101,7 +101,7 @@ TEST(TestHttpEndpointFingerprint, EmptyBody)
     EXPECT_EQ(output.type, DDWAF_OBJ_STRING);
     EXPECT_EQ(attr, object_store::attribute::none);
 
-    std::string_view output_sv{output.stringValue, output.nbEntries};
+    std::string_view output_sv{output.stringValue, static_cast<std::size_t>(output.nbEntries)};
     EXPECT_STRV(output_sv, "http-get-0ede9e60-0ac3796a-");
 
     ddwaf_object_free(&query);
@@ -125,7 +125,7 @@ TEST(TestHttpEndpointFingerprint, EmptyEverything)
     EXPECT_EQ(output.type, DDWAF_OBJ_STRING);
     EXPECT_EQ(attr, object_store::attribute::none);
 
-    std::string_view output_sv{output.stringValue, output.nbEntries};
+    std::string_view output_sv{output.stringValue, static_cast<std::size_t>(output.nbEntries)};
     EXPECT_STRV(output_sv, "http----");
 
     ddwaf_object_free(&query);
@@ -159,7 +159,7 @@ TEST(TestHttpEndpointFingerprint, KeyConsistency)
     EXPECT_EQ(output.type, DDWAF_OBJ_STRING);
     EXPECT_EQ(attr, object_store::attribute::none);
 
-    std::string_view output_sv{output.stringValue, output.nbEntries};
+    std::string_view output_sv{output.stringValue, static_cast<std::size_t>(output.nbEntries)};
     EXPECT_STRV(output_sv, "http-get-0ede9e60-ced401fa-ff07216e");
 
     ddwaf_object_free(&query);
